Use in memory data upload to s3 (#802)
